Refactor package and class names (#2)
