+++ aes-processor/src/main/java/com/ijioio/aes/annotation/processor/exception/EntityIllegalStateException.java
@@ -2,7 +2,7 @@ package com.ijioio.aes.annotation.processor.exception;
 
 import com.ijioio.aes.annotation.processor.MessageContext;
 
-public class EntityTypeIllegalStateException extends ProcessorException {
+public class EntityIllegalStateException extends ProcessorException {
 
 	private static final long serialVersionUID = -3866026527128977914L;
 
@@ -11,7 +11,7 @@ public class EntityTypeIllegalStateException extends ProcessorException {
 	 *
 	 * @param message with the details of the exception
 	 */
-	public EntityTypeIllegalStateException(String message) {
+	public EntityIllegalStateException(String message) {
 		super(message);
 	}
 
@@ -21,7 +21,7 @@ public class EntityTypeIllegalStateException extends ProcessorException {
 	 * @param message with the details of the exception
 	 * @param context defining an exception position
 	 */
-	public EntityTypeIllegalStateException(String message, MessageContext context) {
+	public EntityIllegalStateException(String message, MessageContext context) {
 		super(message, context);
 	}
 }
